split workflow into update and initialize and improve prompt engineering
